Move editor code into lib/editor
